+++ q3/libpoisson.h
@@ -1,2 +1,2 @@
-long factorial(int n);
+long double factorial(long double n);
 long double poisson (long double lambda, long double k);
